requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
basemap==1.3.6
basemap-data==1.3.2
beautifulsoup4==4.11.2
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chromedriver-py==111.0.5563.19
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
dill==0.3.6
discord.py==2.1.0
DiscordUtils==1.3.4
distlib==0.3.6
et-xmlfile==1.1.0
exceptiongroup==1.1.0
facebook-sdk==3.1.0
filelock==3.9.0
Flask==2.2.2
folium==0.14.0
fonttools==4.38.0
frozenlist==1.3.3
google-auth==2.16.0
google-auth-oauthlib==0.8.0
googletrans==3.1.0a0
gspread==5.7.2
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
libretranslatepy==2.1.3
lingua==4.15.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.6.3
multidict==6.0.4
multiprocess==0.70.14
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.26.5
openpyxl==3.0.10
opuslib==3.0.1
outcome==1.2.0
packaging==23.0
pathos==0.3.0
Pillow==9.4.0
platformdirs==3.0.0
playsound==1.3.0
plotly==5.13.0
polib==1.1.1
pox==0.3.2
ppft==1.7.6.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.21
PyExecJS==1.5.1
PyNaCl==1.5.0
pyparsing==3.0.9
pypiwin32==223
pyproj==3.4.1
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.4.3.2
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pyttsx3==2.90
pytz==2022.7.1
pywin32==305
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tenacity==8.2.1
tqdm==4.64.1
translator==0.0.9
translators==5.5.6
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
virtualenv==20.19.0
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
youtube-dl==2021.12.17
zipp==3.12.1
